backend/requirements.txt
# FastAPI and Web Framework
fastapi[standard]>=0.115.7,<0.116.0     # FastAPI 웹 프레임워크
uvicorn[standard]>=0.30.0,<0.31.0       # ASGI 서버 (FastAPI 실행용)

# Database
sqlalchemy>=2.0.23,<2.1.0               # ORM 라이브러리 (DB 모델 관리)
alembic>=1.13.1,<2.0.0                  # DB 마이그레이션 도구
# asyncpg>=0.29.0,<0.30.0                 # 비동기 PostgreSQL 드라이버
# redis>=5.0.1,<6.0.0                     # Redis 클라이언트

# PostgreSQL Driver
psycopg2-binary>=2.9.11                 # PostgreSQL 연동용 Sync 드라이버 (DB 접속)
pgvector>=0.4.0,<0.5.0                  # pgvector extension client library for Python

# Security / Password Handling
passlib[bcrypt]>=1.7.4                  # 비밀번호 해싱(bcrypt) 라이브러리

# Validation / Settings
pydantic>=2.12.4                        # 데이터 검증 및 설정 모델 정의
python-dotenv>=1.2.1                    # .env 환경 변수 로딩 라이브러리

# jwt token
python-jose>=3.3.0                      # JWT 생성/검증 라이브러리 (jose)

# HTTP Client
httpx>=0.24.0                           # 비동기 HTTP 클라이언트

# pandas
pandas>=2.1.0                           # 데이터 처리

# rag
langgraph>=0.3.3
langchain-openai>=0.3.15
langchain>=0.3.30
langchain-community>=0.3.30

# Embedding
torch==2.2.2
numpy<2.0.0
voyageai>=0.3.7

# NLP 관련
konlpy==0.6.0
jpype1==1.5.0  # KoNLPy의 Java 연결을 담당하는 핵심 패키지
scikit-learn>=1.8.0

# ApsScheduler
apscheduler>=3.11.2